Don't report spurious X signals
The idea here is that I want to avoid reporting a signal that is
always X in one of the designs and doesn't exist in the other design.
As a general rule of thumb these signals are meaningless.
Unfortunately Chisel appears to generate a number of these meaningless
signals, so I'm just filtering them out.

I should probably add a "--strict" mode that enforces exact signal
matching...