Don't throw error on missing but constant values
The idea is that some emulators will optimize away constants so much
that they don't even end up in the VCD file at all.  This can be
problematic when trying to diff, but I don't really view it as a wrong
result as it's an invalid signal.

This code keeps track of signals that have never changed and avoids
throwing that error.