Handle having multiple long names with one short name
So I have no idea what's going on here, but VCS appears to sometimes
output some signals with the same short name and expect that this
applies to all the long names.  I fixed this in a somewhat ineffecient
way: just copy the output to all the long names.

For now, I guess it works.